Split SDK executor surface modules
